pyproject.toml
[tool.poetry]
name = "rag-production-system"
version = "1.0.0"
description = "Production-grade RAG system with evaluation, monitoring, and feedback loops"
authors = ["Your Team <team@company.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Async & Database
asyncpg = "^0.29.0"
asyncio = "^3.4.3"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"

# LangChain & RAG
langchain = "^0.1.6"
langchain-aws = "^0.1.0"
langgraph = "^0.0.20"
langchain-community = "^0.0.20"

# AWS SDK
boto3 = "^1.34.0"
botocore = "^1.34.0"

# Text Processing
pypdf2 = "^3.0.0"
beautifulsoup4 = "^4.12.0"
markdown = "^3.5.0"

# Embeddings & Vector Operations
numpy = "^1.26.0"
faiss-cpu = "^1.7.4"  # Optional: for local testing

# Evaluation
deepeval = "^0.20.0"
ragas = "^0.1.0"

# Observability
langsmith = "^0.0.77"

# Utilities
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
tenacity = "^8.2.0"
pydantic-core = "^2.14.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"

# Linting & Formatting
ruff = "^0.1.0"
mypy = "^1.8.0"
black = "^24.0.0"

# Type Stubs
types-boto3 = "^1.0.2"
types-requests = "^2.31.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
coverage = "^7.4.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = """
    --strict-markers
    --cov=app
    --cov=rag
    --cov=graph
    --cov=eval
    --cov-report=term-missing
    --cov-report=xml
    --cov-report=html
"""

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/*" = ["S101"]  # assert usage in tests

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langgraph.*",
    "deepeval.*",
    "ragas.*",
    "boto3.*",
    "botocore.*",
    "pypdf2.*",
]
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | \.tox
  | build
  | dist
  | migrations
)/
'''

[tool.coverage.run]
source = ["app", "rag", "graph", "eval"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
